+++ tests/test_models.py
@@ -8,12 +8,14 @@ from uuid import uuid4
 
 import pytest
 
-from django.contrib.auth.models import User
+from django.contrib import auth
 from django.core.files.uploadedfile import SimpleUploadedFile
 from django.test import TestCase
 
 from user_tasks.exceptions import TaskCanceledException
 from user_tasks.models import UserTaskArtifact, UserTaskStatus
+
+User = auth.get_user_model()
 
 LOGGER = logging.getLogger(__name__)