+++ tests/test_tasks.py
@@ -9,12 +9,14 @@ from uuid import uuid4
 
 from celery import Task, shared_task
 
-from django.contrib.auth.models import User
+from django.contrib import auth
 from django.test import TestCase, override_settings
 from django.utils.timezone import now
 
 from user_tasks.models import UserTaskArtifact, UserTaskStatus
 from user_tasks.tasks import UserTask, UserTaskMixin, purge_old_user_tasks
+
+User = auth.get_user_model()
 
 LOGGER = logging.getLogger(__name__)
 
@@ -42,7 +44,7 @@ class SampleTask(UserTask):  # pylint: disable=abstract-method
     def generate_name(cls, arguments_dict):
         arg1 = arguments_dict['arg1']
         arg2 = arguments_dict['arg2']
-        return 'SampleTask: {}, {}'.format(arg1, arg2)
+        return f'SampleTask: {arg1}, {arg2}'
 
     @staticmethod
     def calculate_total_steps(arguments_dict):